// apps/ui/src/lib/api-client/models/AccessGroupDoor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.5-alpha.202604142009.f75bd474
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * @export
 * @interface AccessGroupDoor
 */
export interface AccessGroupDoor {
    /**
     * Assignment ID
     * @type {string}
     * @memberof AccessGroupDoor
     */
    id: string;
    /**
     * Group ID
     * @type {string}
     * @memberof AccessGroupDoor
     */
    groupId: string;
    /**
     * Door ID
     * @type {string}
     * @memberof AccessGroupDoor
     */
    doorId: string;
    /**
     * Created timestamp (ISO 8601)
     * @type {string}
     * @memberof AccessGroupDoor
     */
    createdAt?: string;
    /**
     * Updated timestamp (ISO 8601)
     * @type {string}
     * @memberof AccessGroupDoor
     */
    updatedAt?: string;
}

/**
 * Check if a given object implements the AccessGroupDoor interface.
 */
export function instanceOfAccessGroupDoor(value: object): value is AccessGroupDoor {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('groupId' in value) || value['groupId'] === undefined) return false;
    if (!('doorId' in value) || value['doorId'] === undefined) return false;
    return true;
}

export function AccessGroupDoorFromJSON(json: any): AccessGroupDoor {
    return AccessGroupDoorFromJSONTyped(json, false);
}

export function AccessGroupDoorFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccessGroupDoor {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'groupId': json['groupId'],
        'doorId': json['doorId'],
        'createdAt': json['createdAt'] == null ? undefined : json['createdAt'],
        'updatedAt': json['updatedAt'] == null ? undefined : json['updatedAt'],
    };
}

export function AccessGroupDoorToJSON(json: any): AccessGroupDoor {
    return AccessGroupDoorToJSONTyped(json, false);
}

export function AccessGroupDoorToJSONTyped(value?: AccessGroupDoor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'groupId': value['groupId'],
        'doorId': value['doorId'],
        'createdAt': value['createdAt'],
        'updatedAt': value['updatedAt'],
    };
}